Use the k8s dynamic client to discover source CRDs
The previous implementation was hard-coded to only read a specific
kind of Object: endpoint.DNSEndpoint. This implementation uses the k8s
client-go to process any k8s resource that "quacks" like a
DNSEndpoint, i.e., any Object whose unstructured representation can be
converted into a valid DNSEndpoint. I can only guess at this point,
but I think that this was the original idea behind the crd-source-kind
flag, but over time it lost the ability to do this.

Now a user can add the fields from DNSEndpoint (both Spec and Status)
to their own CRDs and have external-dns use their CRs as a Source.

The --crd-source-kind flag never actually worked, so I've renamed it
to --crd-source-resource. It now takes the resource name, so where
--crd-source-kind used to take something like "DNSEndpoint",
--crd-source-resource now takes something like "dnsendpoints". The
kubectl api-resources command shows the mapping from kinds to
resources.